refactor(InputValidator): 유효성 검사 메서드 수정 - 기존의 유효성 검사는 잘못된 입력을 했을 때 false의 부울(Bool)값을 반환하도록 하였지만 IllegalArgumentException 예외를 던지도록 변경함.
